requirements.txt
tweepy==3.9.0
python-dotenv==0.14.0
Wikidata==0.7.0
pillow==8.2.0
unidecode==1.2.0
wget==3.2